+++ CHANGES.rst
@@ -11,6 +11,7 @@ Improvements
 ^^^^^^^^^^^^
 
 - Add a privacy policy page linked from the footer (:issue:`1415`)
+- *Terms & Conditions* can now link to an external URL
 
 Bugfixes
 ^^^^^^^^