+++ CHANGES.rst
@@ -13,6 +13,8 @@ Improvements
 - Add a privacy policy page linked from the footer (:issue:`1415`)
 - *Terms & Conditions* can now link to an external URL
 - Show a warning to all admins if Celery is not running or outdated
+- Add registration ID placeholder for badges (:issue:`3370`, thanks
+  :user:`bpedersen2`)
 
 Bugfixes
 ^^^^^^^^